Do not anticipate MatchType shared class publication
Keep all MetricTagsMatch fields in a single class
