+++ src/main/java/examples/MetricsExamples.java
@@ -21,8 +21,8 @@ import io.vertx.core.http.HttpClientOptions;
 import io.vertx.core.json.JsonObject;
 import io.vertx.docgen.Source;
 import io.vertx.ext.hawkular.AuthenticationOptions;
-import io.vertx.ext.hawkular.MatchType;
 import io.vertx.ext.hawkular.MetricTagsMatch;
+import io.vertx.ext.hawkular.MetricTagsMatch.MatchType;
 import io.vertx.ext.hawkular.VertxHawkularOptions;
 
 import java.util.Arrays;
