--- src/main/java/io/vertx/ext/hawkular/Match.java
@@ -1,100 +1,0 @@
-/*
- * Copyright 2017 Red Hat, Inc.
- *
- *  All rights reserved. This program and the accompanying materials
- *  are made available under the terms of the Eclipse Public License v1.0
- *  and Apache License v2.0 which accompanies this distribution.
- *
- *  The Eclipse Public License is available at
- *  http://www.eclipse.org/legal/epl-v10.html
- *
- *  The Apache License v2.0 is available at
- *  http://www.opensource.org/licenses/apache2.0.php
- *
- *  You may elect to redistribute this code under either of these licenses.
- */
-
-package io.vertx.ext.hawkular;
-
-import io.vertx.codegen.annotations.DataObject;
-import io.vertx.core.json.JsonObject;
-
-/**
- * A match for a value.
- *
- * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
- */
-@DataObject
-public class Match {
-
-  /**
-   * The default value : {@link MatchType#EQUALS}
-   */
-  public static final MatchType DEFAULT_TYPE = MatchType.EQUALS;
-
-  private String value;
-  private MatchType type;
-
-  /**
-   * Default constructor
-   */
-  public Match() {
-    type = DEFAULT_TYPE;
-  }
-
-  /**
-   * Copy constructor
-   *
-   * @param other The other {@link Match} to copy when creating this
-   */
-  public Match(Match other) {
-    value = other.value;
-    type = other.type;
-  }
-
-  /**
-   * Create an instance from a {@link io.vertx.core.json.JsonObject}
-   *
-   * @param json the JsonObject to create it from
-   */
-  public Match(JsonObject json) {
-    value = json.getString("value");
-    type = MatchType.valueOf(json.getString("type", DEFAULT_TYPE.name()));
-  }
-
-  /**
-   * @return the matched value
-   */
-  public String getValue() {
-    return value;
-  }
-
-  /**
-   * Set the matched value.
-   *
-   * @param value the value to match
-   * @return a reference to this, so the API can be used fluently
-   */
-  public Match setValue(String value) {
-    this.value = value;
-    return this;
-  }
-
-  /**
-   * @return the matcher type
-   */
-  public MatchType getType() {
-    return type;
-  }
-
-  /**
-   * Set the type of matching to apply.
-   *
-   * @param type the matcher type
-   * @return a reference to this, so the API can be used fluently
-   */
-  public Match setType(MatchType type) {
-    this.type = type;
-    return this;
-  }
-}
