--- src/main/java/io/vertx/ext/hawkular/MatchType.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2017 Red Hat, Inc.
- *
- *  All rights reserved. This program and the accompanying materials
- *  are made available under the terms of the Eclipse Public License v1.0
- *  and Apache License v2.0 which accompanies this distribution.
- *
- *  The Eclipse Public License is available at
- *  http://www.eclipse.org/legal/epl-v10.html
- *
- *  The Apache License v2.0 is available at
- *  http://www.opensource.org/licenses/apache2.0.php
- *
- *  You may elect to redistribute this code under either of these licenses.
- */
-
-package io.vertx.ext.hawkular;
-
-/**
- * The type of match.
- *
- * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
- */
-public enum MatchType {
-
-  EQUALS,
-
-  REGEX
-
-}
